Fixed trailer issue
I have changed the foreach loop that I used to loop through videosArr because, I was always getting the latest Trailer result instead of the first result matching the if condition.
By changing the loop with a for of loop, I can now break the loop after finding the first matching result.